test: separate integration from unit
