+++ app/database/mapper.py
@@ -7,8 +7,8 @@ from app.domain.entities.sms_type import SmsType
 from app.domain.entities.message import Message
 from app.domain.entities.phone_number import PhoneNumber
 from app.domain.entities.sms_status import SmsDeliveryStatus
-from .sms_model import Sms as SmsModel
-from .sms_response_model import SmsResponse as SmsResponseModel
+from app.database.models.sms_model import Sms as SmsModel
+from app.database.models.sms_response_model import SmsResponse as SmsResponseModel
 
 
 def map_sms_model_to_entity(model: SmsModel) -> Sms:
